fixed some tests
