+++ app/src/androidTest/java/ch/epfl/skysync/ConfirmFlightScreenTest.kt
@@ -38,7 +38,7 @@ class ConfirmFlightScreenTest {
       navController = TestNavHostController(LocalContext.current)
       navController.navigatorProvider.addNavigator(ComposeNavigator())
       NavHost(navController = navController, startDestination = Route.MAIN) {
-        homeGraph(repository, navController, dbs.admin1.id)
+        homeGraph(repository, navController, dbs.crew1.id)
       }
     }
     composeTestRule.waitUntil {
