+++ app/src/androidTest/java/ch/epfl/skysync/components/calendar/AvailabiltyCalendarTest.kt
@@ -1,4 +1,4 @@
-package ch.epfl.skysync.screens.calendar
+package ch.epfl.skysync.components.calendar
 
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.test.assertIsDisplayed
@@ -44,7 +44,7 @@ class AvailabiltyCalendarTest {
       navController = TestNavHostController(LocalContext.current)
       navController.navigatorProvider.addNavigator(ComposeNavigator())
       NavHost(navController = navController, startDestination = Route.MAIN) {
-        homeGraph(repository, navController, dbs.admin1.id)
+        homeGraph(repository, navController, dbs.crew1.id)
       }
     }
     composeTestRule.waitUntil {
