+++ app/src/androidTest/java/ch/epfl/skysync/components/calendar/CalendarTopBarTest.kt
@@ -36,7 +36,7 @@ class CalendarTopBarTest {
       navController = TestNavHostController(LocalContext.current)
       navController.navigatorProvider.addNavigator(ComposeNavigator())
       NavHost(navController = navController, startDestination = Route.MAIN) {
-        homeGraph(repository, navController, dbs.admin1.id)
+        homeGraph(repository, navController, dbs.crew1.id)
       }
     }
     composeTestRule.waitUntil {
