+++ app/src/androidTest/java/ch/epfl/skysync/components/calendar/FlightCalendarTest.kt
@@ -41,7 +41,7 @@ class FlightCalendarTest {
       navController = TestNavHostController(LocalContext.current)
       navController.navigatorProvider.addNavigator(ComposeNavigator())
       NavHost(navController = navController, startDestination = Route.MAIN) {
-        homeGraph(repository, navController, dbs.admin1.id)
+        homeGraph(repository, navController, dbs.crew1.id)
       }
     }
     composeTestRule.waitUntil {
