+++ app/src/androidTest/java/ch/epfl/skysync/screens/home/UpcomingFlightsTests.kt
@@ -4,10 +4,10 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
+import ch.epfl.skysync.components.UpcomingFlights
 import ch.epfl.skysync.models.calendar.TimeSlot
 import ch.epfl.skysync.models.flight.FlightType
 import ch.epfl.skysync.models.flight.PlannedFlight
-import ch.epfl.skysync.screens.UpcomingFlights
 import ch.epfl.skysync.ui.theme.lightOrange
 import java.time.LocalDate
 import org.junit.Assert.assertTrue
